Convert to uint64_t where applicable

+++ src/builder.cc
@@ -652,7 +652,7 @@ void RTLILBuilder::add_memory_init(
 		processed += length;
 	}
 
-	if (processed < data.size()) {
+	if (processed < (uint64_t)data.size()) {
 		log_assert((bit_offset + processed) % mem->width == 0);
 		uint64_t length = ((((uint64_t)data.size()) - processed) / mem->width) * mem->width;
 		emit_meminit_cell(mem, (bit_offset + processed) / mem->width, big_endian,
@@ -660,9 +660,9 @@ void RTLILBuilder::add_memory_init(
 		processed += length;
 	}
 
-	if (processed < data.size()) {
-		uint64_t length = data.size() - processed;
-		log_assert(length < mem->width);
+	if (processed < (uint64_t)data.size()) {
+		uint64_t length = (uint64_t)data.size() - processed;
+		log_assert(length < (uint64_t)mem->width);
 		Const data1, mask1;
 		data1.append(data.extract(0, length));
 		data1.append(Const(Sx, mem->width - length));
@@ -672,7 +672,7 @@ void RTLILBuilder::add_memory_init(
 		processed += length;
 	}
 
-	log_assert(processed == data.size());
+	log_assert(processed == (uint64_t)data.size());
 }
 
 SigSpec RTLILBuilder::add_placeholder_signal(
